Uploading progress up till now, work done off of github.
Moved passwds and keys out to config.py and added to .gitignore. Current status of project asks for CLI user input, looks through the content of the website for a few basic patterns indicating two styles of NPR Music sites, and rips a list of songs out of those sites if possible. It then checks the songs and artists against the lastfm for recommended spellcheck fixes, and asks for confirmation before scrobbling the list.